Committing to string-based source code, adding primitive real numbers.
Free variables lead to dynamic linking which is very expressive and potentially useful. I was concerned about code not having stable meaning due to different bindings, but I think now that it's a technique to be used rather than something to be avoided.

Numbers are too important to be left to accelerators, particularly "real" numbers which you weren't going to accelerate anyway (implementing floating point, with all its quirks, in ABC? Really?).